requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
attrs==19.1.0
cffi==1.12.3
Click==7.0
construct==2.9.45
cryptography==2.6.1
ifaddr==0.1.6
netifaces==0.10.9
pycparser==2.19
python-miio==0.4.5
pytz==2019.1
six==1.12.0
tqdm==4.31.1
zeroconf==0.22.0
